+++ pages/add_account/index.tsx
@@ -1,12 +1,26 @@
-import { LoginLayout } from "@layouts/LoginLayout";
-import { ReactElement } from "react";
-
-
+import LoginForm from '@components/Auth/LoginForm';
+import { LoginLayout } from '@layouts/LoginLayout';
+import { sendRequest } from '@requests/request';
+import { ReactElement, useCallback } from 'react';
 
 function AddAccount() {
-  return (
-    <>Add account</>
-  )
+  const signIn = useCallback(
+    (org: string, login: string, password: string) =>
+      sendRequest<
+        {
+          organization: string;
+          login: string;
+          password: string;
+        },
+        boolean
+      >('/auth/add_account', 'POST', {
+        organization: org,
+        login,
+        password,
+      }).then((res) => res.response),
+    []
+  );
+  return <LoginForm signIn={signIn} />;
 }
 
 AddAccount.getLayout = (page: ReactElement) => {
