+++ pages/api/auth/add_account.ts
@@ -1,0 +1,14 @@
+import { fetchWrapper } from '@utils/fetchWrapper';
+import { NextApiRequest, NextApiResponse } from 'next';
+
+export default async function AddAccount(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  await fetchWrapper({
+    req: req,
+    res: res,
+    url: 'api/add_account',
+    method: 'POST',
+  });
+}
